Do not display metadata in source trees
Move metadata detection to DataSourceUtils, since it doesn't belong with NetworkTable utilities
