+++ api/src/test/java/edu/wpi/first/shuffleboard/api/util/NetworkTableUtilsTest.java
@@ -2,7 +2,6 @@ package edu.wpi.first.shuffleboard.api.util;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
@@ -10,28 +9,6 @@ import java.util.stream.Stream;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class NetworkTableUtilsTest extends UtilityClassTest<NetworkTableUtils> {
-
-
-  @ParameterizedTest
-  @CsvSource({"false, /key",
-                 "false, /metadata",
-                 "false, /~metadata",
-                 "false, /metadata~",
-                 "true, /~metadata~",
-                 "false, /METADATA",
-                 "false, /~METADATA",
-                 "false, /METADATA~",
-                 "true, /~METADATA~",
-                 "true, /.metadata",
-                 "true, /~METADATA~/someOtherValue",
-                 "true, /.metadata/someOtherValue",
-                 "false, /my.key.with.dots",
-                 "false, /my~key~with~tildes~",
-                 "false, /~my~keywithtildes",
-                 "false, /~metadata~with~tildes~"})
-  public void isMetaDataTest(boolean expectedResult, String key) {
-    assertEquals(expectedResult, NetworkTableUtils.isMetadata(key));
-  }
 
   private static Stream<Arguments> concatArguments() {
     return Stream.of(
