+++ app/src/main/java/edu/wpi/first/shuffleboard/app/components/DashboardTab.java
@@ -5,12 +5,12 @@ import edu.wpi.first.shuffleboard.api.components.ExtendedPropertySheet;
 import edu.wpi.first.shuffleboard.api.data.DataTypes;
 import edu.wpi.first.shuffleboard.api.prefs.FlushableProperty;
 import edu.wpi.first.shuffleboard.api.sources.DataSource;
+import edu.wpi.first.shuffleboard.api.sources.DataSourceUtils;
 import edu.wpi.first.shuffleboard.api.sources.SourceType;
 import edu.wpi.first.shuffleboard.api.sources.SourceTypes;
 import edu.wpi.first.shuffleboard.api.tab.TabInfo;
 import edu.wpi.first.shuffleboard.api.util.Debouncer;
 import edu.wpi.first.shuffleboard.api.util.FxUtils;
-import edu.wpi.first.shuffleboard.api.util.NetworkTableUtils;
 import edu.wpi.first.shuffleboard.api.util.TypeUtils;
 import edu.wpi.first.shuffleboard.api.widget.Component;
 import edu.wpi.first.shuffleboard.api.widget.ComponentContainer;
@@ -251,7 +251,7 @@ public class DashboardTab extends Tab implements HandledTab, Populatable {
         && isAutoPopulate()
         && !getSourcePrefix().isEmpty()
         && type.dataTypeForSource(DataTypes.getDefault(), sourceId) != DataTypes.Map
-        && !NetworkTableUtils.isMetadata(sourceId)
+        && DataSourceUtils.isNotMetadata(sourceId)
         && (name.startsWith(getSourcePrefix()) || sourceId.startsWith(getSourcePrefix()));
   }
